requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
bcrypt==4.3.0
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
boto3==1.34.60
